requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.2
Django==4.0.3
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.0.0
